requirements.txt
langchain
openai==0.28.1
streamlit
python-dotenv
bs4
chromadb
tiktoken
streamlit-aggrid